create separate folder for hf inside pennylane
